Only change render state for flat textures without color key
Color keys and linear filtering doesn't mix. The transparent parts of the
texture need to have the same color as their neighbors, and alpha
transparency instead for linear filtering to work, and I'm not quite sure
how to achive this. But only changing the render state for a select few
textures where it's save already has some noticably positive effects.