Fix the video upscaling problem properly by using a decoy backbuffer
Instead of the real backbuffer, the game is given 640x480-surface, where
it can do whatever it wants, and on UnLock, this surface is simply upscaled
to fit the real resolution.

This is a very clean solution which does not require any heuristics. The
only potential disadvantage could be a situation where the game would have
done something reasonable with a bigger backbuffer, but that's unlikely and
the result would still look as good as the original.