launcher: start ff8 suspended and resume after the dll is injected
Mostly done so that a debugger can stop ff8 with a breakpoint in the
launcher and attach to ff8 at that point. Though with ddraw blocking the
debugger gui, it's mostly useless.

It should prevent a theoretical race condition where ff8 makes use of dx
before the dll can hook the calls.