Apply the DX 0.5 offset which the game designers forgot
In DX, the screen spans from -0.5, -0.5 to resX-0.5, resY-0.5 instead of
0, 0 to resX, resY. This is important if one uses antialiasing, because a
rect from 0, 0 to resX, resY does not cover the whole screen, and visual
artifacts at the top and left edges occur. The game developers apparently
didn't know this, and the wrapper failed to take this into account when
scaling up, too. Just fixing the wrapper isn't enough, it has to compensate
for the original coordinates, too.