Prepare for scale factors which make the game bigger than the screen
Since the original game has black borders at the top and bottom, it has
black borders on all sides with a widescreen display. To solve this, the
scale factor needs to be bigger, which would result in viewports bigger
than the screen. DX does not allow this, so some adjustments are needed.